drivers: power: max17050: change voltage_max_design to uint32_t to prevent truncation
Reviewed-on: http://gerrit.mot.com/812263
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
SLTApproved: Slta Waiver <sltawvr@motorola.com>
SME-Granted: SME Approvals Granted

+++ nuttx/drivers/power/max17050_config.h
@@ -57,7 +57,7 @@ struct max17050_config {
     uint16_t capacity;
     uint16_t vf_fullcap;
     uint16_t ichg_term;
-    uint16_t voltage_max_design; // in uV units
+    uint32_t voltage_max_design; // in uV units
     int cell_data[MAX17050_CELL_DATA_SIZE];
 };
